add encoder and reconfig createUser
