+++ src/main/kotlin/com/gamerly/projectgamerly/service/UsuarioService.kt
@@ -84,52 +84,52 @@ class UsuarioService {
         return usuariosFiltrados.map{usuario -> UsuarioBusquedaDto(usuario) }
     }
 
-    fun login(credenciales: CredencialesDTO): UsuarioLoginDTO {
-        val usuarioCrendecial = Usuario().apply {
-            email = credenciales.email
-            password = credenciales.password
-            tokenNotificaciones = credenciales.tokenNotificaciones
-        }
-        val usuario = usuarioRepository.findByEmail(usuarioCrendecial.email)
+//    fun login(credenciales: CredencialesDTO): UsuarioLoginDTO {
+//        val usuarioCrendecial = Usuario().apply {
+//            email = credenciales.email
+//            password = credenciales.password
+//            tokenNotificaciones = credenciales.tokenNotificaciones
+//        }
+//        val usuario = usuarioRepository.findByEmail(usuarioCrendecial.email)
+//
+//        if (usuario.isEmpty) {
+//            throw CredencialesInvalidas("Credenciales invalidas")
+//        }
+//
+//        val usuarioEncontrado = usuario.get()
+//
+//        if (usuarioEncontrado.password != usuarioCrendecial.password) {
+//            throw CredencialesInvalidas("Credenciales invalidas")
+//        }
+//
+//        usuarioEncontrado.tokenNotificaciones = credenciales.tokenNotificaciones
+//        usuarioRepository.save(usuarioEncontrado)
+//
+//        return UsuarioLoginDTO.from(usuarioEncontrado);
+//    }
 
-        if (usuario.isEmpty) {
-            throw CredencialesInvalidas("Credenciales invalidas")
-        }
-
-        val usuarioEncontrado = usuario.get()
-
-        if (usuarioEncontrado.password != usuarioCrendecial.password) {
-            throw CredencialesInvalidas("Credenciales invalidas")
-        }
-
-        usuarioEncontrado.tokenNotificaciones = credenciales.tokenNotificaciones
-        usuarioRepository.save(usuarioEncontrado)
-
-        return UsuarioLoginDTO.from(usuarioEncontrado);
-    }
-
-    fun crearUsuario(user: UsuarioCreacionDTO): Usuario {
-        val usuario = usuarioRepository.findByEmail(user.email).orElse(null)
-
-        if(usuario != null){
-            throw InvalidEmail("Ya existe un usuario registrado con ese email")
-        }
-
-        val usuarioRegistro = Usuario().apply {
-            nombre = user.nombre
-            fechaDeNacimiento = LocalDate.parse(
-                user.fechaNacimiento,
-                DateTimeFormatter.ofPattern("dd/MM/yyyy")
-            )
-            email = user.email
-            password = user.password
-            discord = user.discord
-            nacionalidad = user.nacionalidad
-            foto = "https://i.ibb.co/HG1GTNR/avatar.png"
-
-        }
-        return usuarioRepository.save(usuarioRegistro)
-    }
+//    fun crearUsuario(user: UsuarioCreacionDTO): Usuario {
+//        val usuario = usuarioRepository.findByEmail(user.email).orElse(null)
+//
+//        if(usuario != null){
+//            throw InvalidEmail("Ya existe un usuario registrado con ese email")
+//        }
+//
+//        val usuarioRegistro = Usuario().apply {
+//            nombre = user.nombre
+//            fechaDeNacimiento = LocalDate.parse(
+//                user.fechaNacimiento,
+//                DateTimeFormatter.ofPattern("dd/MM/yyyy")
+//            )
+//            email = user.email
+//            password = user.password
+//            discord = user.discord
+//            nacionalidad = user.nacionalidad
+//            foto = "https://i.ibb.co/HG1GTNR/avatar.png"
+//
+//        }
+//        return usuarioRepository.save(usuarioRegistro)
+//    }
 
     fun editarUsuario(idUsuario: Long, usuarioEditado: UsuarioEditarDTO): UsuarioDetalleDTO {
         val usuario = usuarioRepository.findById(idUsuario)
